Automatically refresh menu list after modify the data in config/plugin
